Cleaning up more
