--- data.json
@@ -1,1 +1,0 @@
-[{"email":"family.man.123@finau-family.org","firstName":"Tony","lastName":"Finau","pass":"$2a$10$LrIhvULT8pFa7ZvYz2IzKelqLeqaNk1kcdKLmspPpALzPtynbQwxu","pid":"05ae2282-6781-457c-8271-32d7c52e8705"},{"email":"kbradley@test.com","firstName":"Keegan","lastName":"Bradley","pass":"$2a$10$EfSxt6Ps3OlUPBlyq/YyXOlv.jqZKPgI9KC8PbtvNc/jZ7a/TPE1e","pid":"163630f8-ba94-4030-b716-cceedd1be8dc"},{"email":"wyndham@foo.com","firstName":"Wyndham","lastName":"Clark","pass":"$2a$10$EYpNp82lgdqYBEFNzbF0U.85dTmv25SGx5.0cIEy9epSMebM4gTJK","pid":"17f37e32-7162-452c-94b3-f274b82561e5"},{"email":"Schauffele@gmail.com","firstName":"Xander ","lastName":"Schauffele","pass":"$2a$10$eYRNVssqTGLyQnUqozk.0OCtlTq2oqu2QLQhIODG9z9.AM1qWGLBK","pid":"247783ba-6741-42ac-a200-d7537514192f"},{"email":"suh.dude@gmail.com","firstName":"Justin","lastName":"Suh","pass":"$2a$10$U3SM6RgguOwP8.KGNFsq1.O4tnnY5W.abV1frU8gtyHnoeCmeTqw.","pid":"281f1e7a-7a8b-46e2-995d-872e8a5e314f"},{"email":"tw@tw.com","firstName":"Tiger","lastName":"Woods","pass":"$2a$10$Q.lNzH4qspV.7VQkNsVIyedxjUgb6p5F3oM5RjoXn6DK6WGSxaWMW","pid":"316f4cd5-2bfc-40be-89e4-9ae6ce2dc940"},{"email":"slow-golfer@pga.com","firstName":"Patrick","lastName":"Cantlay","pass":"$2a$10$zMsHeMRNz.PGdloRgl0Z7OlIeFoFpgmjxOqzC6cKn7wd0ohokFekO","pid":"39545dd5-13d5-4851-9d94-879df42a67dc"},{"email":"elliot.scribner@couchbase.com","firstName":"Test","lastName":"Guy","pass":"$2a$10$1HneyN.kvavXpvlL1V1H6evZDeTmvQWEeOu42V3QLiZBt57BNPKDK","pid":"49469bbf-4e04-49ad-b2c3-eaa9653a500e"},{"email":"vik.snap@foo.com","firstName":"Viktor","lastName":"Hovland","pass":"$2a$10$ehazoqnnQTQHcjuhNtpOMevJl620NDmtQWVWhJovhnNBA1iVndRqG","pid":"4e9c6c5b-3676-41b3-a99b-b7430f694430"},{"email":"neurotic-golfer@gmail.com","firstName":"Matt","lastName":"Fitzpatrick","pass":"$2a$10$L1bwfo/IGJa44J.PHbKdKe0GSUPdbXjJ7e0gi9RcpoZ2Rl5DCLwTG","pid":"5317d039-9640-444a-9dce-33e318457384"},{"email":"jd.miller@test.com","firstName":"John","lastName":"Daly","pass":"$2a$10$rL92CBFXNgM3M5gZ3IBxOec4LP3zOgzxCminYUaRa5s7Jq6iSl4li","pid":"5fb6dc67-f6bf-4823-85e8-841628ea7f25"},{"email":"jt@pga.com","firstName":"Justin","lastName":"Thomas","pass":"$2a$10$qiHRXFWo5uCpnPVHYznnK.zynzNqVvjT3GNW6UlNGbtGbEQpqPAoC","pid":"770b0616-6a7f-44b4-8d5f-9a4e2db7696b"},{"email":"adamSchenksTheBall@pga.org","firstName":"Adam","lastName":"Schenk","pass":"$2a$10$Z/j4B0l6a5LIdekCZDOlbegQQRubdoxJqu1s6OmYot5qTJubDG/8i","pid":"79b7aace-cc12-4128-8fad-18b4d2f0b84f"},{"email":"sam@sam.com","firstName":"Sam","lastName":"Bennett","pass":"$2a$10$YLguaL9.1O2bql.52GIuhOMv0/svXWO5RjlohMihTdQSdFrpKw4V6","pid":"7df6cd99-2dd1-41d7-92c0-ca16a205e62d"},{"email":"masterschamp@yahoo.com","firstName":"Jon","lastName":"Rahm","pass":"$2a$10$0UUb4.leMISLClCwPqip/.RJT7KkpGOzxlwRQ549t3JJOVi3oeP4y","pid":"afb951aa-134d-4459-8a7a-1d6bfc936d53"},{"email":"richard.fowler.the.third@longemailprovider.com","firstName":"Rickie","lastName":"Fowler","pass":"$2a$10$ZDmfqcMvslvbO.vXtsIOBuTglVi4oky9JYVOfWp9EjvYJx10NAcAi","pid":"b952d65a-e212-44e5-8938-09277c2ff16f"},{"email":"123@456.com","firstName":"Test","lastName":"Guy","pass":"$2a$10$VfjXe.1GDhnYFDvoQpBrv.BWC3DhImKSIR9DyBtHjg2uW3jEuQX8q","pid":"b9729458-fbb4-4038-9e8b-5abd6c48ead1"},{"email":"livgolfer216@PIF.com","firstName":"Cam","lastName":"Smith","pass":"$2a$10$JCutYzzu2u5TWzOS8gKe6e6cImM/LNOnivtxJ2q3c5fQyW5cjLEsu","pid":"bc60eca2-3526-4c36-bdaa-add7ea7b6791"},{"email":"elliot.scribner@couchbase.com","firstName":"Elliot","lastName":"Scribner","pass":"$2a$10$FHzGPxbzcplZcyJpAJA8/eWG9ssnf6KlDWo.USH4VtP9fN.d77QDi","pid":"c05d14dc-9b97-49e3-bbe8-b799ef36c37d"},{"email":"brooks@liv.com","firstName":"Brooks","lastName":"Koepka","pass":"$2a$10$IdBYZhjOwv1I13OkAgucCuWLuwZqDHXkQn9V3HjmL9uojYplL7sTi","pid":"caa47bc8-ac7a-467d-9144-0807e8d0db7c"},{"email":"denny@test.com","firstName":"Denny","lastName":"McCarthy","pass":"$2a$10$LRXwbSNSYdYQuumL.zsR3.C5jqR4uq2So.z/mOsRV2caCYjV2VotG","pid":"ced2bab1-63e5-4624-ae1c-42999c609e06"},{"email":"jordan@speith.com","firstName":"Jordan","lastName":"Speith","pass":"$2a$10$f8qizdNB5MFLJ6hHBF3xLOhYv6Wy9BEHldfq4cbcC.4MHloEVVehq","pid":"eaa7cfa0-56ee-446c-ae4b-695c4f1c04d2"},{"email":"c.morikowa@gmail.com","firstName":"Colin","lastName":"Morikowa","pass":"$2a$10$/KnZxZv4pKf4ai3VQvFOIeegEoDsvOK5THYka.1/h35k5tNofqAEu","pid":"f62fe131-8467-4ebd-8813-0ecd64412136"},{"email":"masterschampbeforejon@yahoo.com","firstName":"Scottie","lastName":"Scheffler","pass":"$2a$10$r8ASTuAHTLVftLQmkeUtQ.N6EF87T9xLhhh5r2FBKQo8wV5KgncsC","pid":"f79abfc8-e640-402b-9718-b92f7410fb3d"},{"email":"washed@liv.com","firstName":"Phil","lastName":"Mickleson","pass":"$2a$10$nh02WdjmP8Gj9XKPwOCV.u.VIg9Imj.wg3HuS2rZ/iTWul5/zOxIO","pid":"fcb0f020-f1de-49e4-8938-60bc29b0fc66"}]
